% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bluesky_archive_at_locs}
\alias{bluesky_archive_at_locs}
\title{bluesky_archive_at_locs}
\usage{
bluesky_archive_at_locs(locs, path = "./data/bluesky/archive/")
}
\arguments{
\item{locs}{A SpatialPointsDataFrame of locations and dates}

\item{path}{The path to bluesky archive data (default is
"./data/bluesky/archive/)}
}
\value{
The data from \emph{locs} with BlueSky PM2.5 appended (as PM25_bluesky)
}
\description{
For a given spatial points data frame, assign bluesky archive values for all
of the needed dates
}
\examples{
bluesky <- bluesky_archive_at_locs(locations)
}
