% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{create_airnow_variograms}
\alias{create_airnow_variograms}
\title{create_airnow_variograms}
\usage{
create_airnow_variograms(df, cutoff = NULL)
}
\arguments{
\item{df}{A SpatialPointsDataFrame from \code{\link{recast_monitors}}. Can
include AirNow and/or AIRSIS data.}

\item{cutoff}{numeric A distance passed to \code{\link[gstat]{variogram}}}
}
\value{
A list of variograms, one for each date in the input data set
}
\description{
Calculates daily spatial variograms for kriging interpolation on monitor data
}
\examples{
an_vg <- create_airnow_variograms(an_ws)
}
