% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{create_purpleair_archive}
\alias{create_purpleair_archive}
\title{create_purpleair_archive}
\usage{
create_purpleair_archive(dt1, dt2, states = "CA", pas = NULL)
}
\arguments{
\item{dt1}{Date The earliest date to search}

\item{dt2}{Date The latest date to search}

\item{states}{character A vector of two-character state codes}

\item{pas}{An optional pas object that contains the list of sites to search
for}
}
\value{
A pas-like object of PurpleAir data that can be used in further
  processing
}
\description{
Download and process PurpleAir for given states over a given daterange.
}
\examples{
pa_data <- create_purpleair_archive(dt1, dt2, "CA")
}
