% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{create_purpleair_variograms}
\alias{create_purpleair_variograms}
\title{create_purpleair_variograms}
\usage{
create_purpleair_variograms(df, cutoff = NULL)
}
\arguments{
\item{df}{A cleaned SpatialPointsDataFrame of PurpleAir data, such as from
\code{\link{purpleair_clean_spatial_outliers}}}

\item{cutoff}{numeric A cutoff distance in meters that is passed to
\code{\link[gstat]{vgm}}}
}
\value{
A list containing daily variograms and model fits
}
\description{
Create binned variograms using a cleaned PurpleAir data set, one day at a
time
}
\examples{
pa_vg <- create_purpleair_variograms(pa_clean, cutoff = 100000)
}
