% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{daily_map}
\alias{daily_map}
\title{Title}
\usage{
daily_map(dt, data_path, validation_path, monitor_path, state = NULL)
}
\arguments{
\item{dt}{}

\item{data_path}{}

\item{validation_path}{}

\item{monitor_path}{}

\item{state}{}
}
\description{
Title
}
