% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop_model}
\alias{develop_model}
\title{Title}
\usage{
develop_model(dt1, dt2, paths, test_extent = NULL, pa_nmax = 100, seed = 1977)
}
\arguments{
\item{dt1}{}

\item{dt2}{}

\item{paths}{}

\item{test_extent}{}

\item{pa_nmax}{}

\item{seed}{}
}
\description{
Title
}
