% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop_model_h}
\alias{develop_model_h}
\title{Title}
\usage{
develop_model_h(
  dt1,
  dt2,
  states,
  pa_cutoff = 1e+05,
  seed = 1977,
  pa_data = NULL,
  maiac_path = "./data/MAIAC/"
)
}
\arguments{
\item{dt1}{}

\item{dt2}{}

\item{states}{}

\item{pa_cutoff}{}

\item{seed}{}

\item{pa_data}{}

\item{maiac_path}{}
}
\description{
Title
}
