% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{get_airsis_daterange}
\alias{get_airsis_daterange}
\title{get_airsis_daterange}
\usage{
get_airsis_daterange(start, end, states)
}
\arguments{
\item{start}{Date The earliest date to search}

\item{end}{Date The latest date to search. Must be in the same calendar year
as \code{start}.}

\item{states}{character The states to include as a vector of two-character
state abbreviations}
}
\value{
AIRSIS data for the specified dates and locations as A SpatialPointsDataFrame
}
\description{
NOTE!! As of 2024-09-05, the most recent airsis data in the AirFire archive was 2022-09-19
}
\details{
For the specified date range and US states, downloads AIRSIS data.
}
\examples{
dt1 <- as.Date("2018-11-01")
  dt2 <- as.Date("2018-11-30")
  an_ws <- get_airsis_daterange(dt1, dt2, c("CA", "NV"))
}
