% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{get_monitor_daterange}
\alias{get_monitor_daterange}
\title{get_monitor_daterange}
\usage{
get_monitor_daterange(start, end, states, mon_source = "airnow")
}
\arguments{
\item{start}{Date The earliest date to search}

\item{end}{Date The latest date to search. Must be in the same calendar year as \code{start}.}

\item{states}{character The states to include as a vector of two-character state abbreviations}

\item{mon_source}{character The monitor source type among "airnow", "airsis", or "wrcc"}
}
\value{
monitoring data for the specified dates and locations as SpatialPointsDataFrame
}
\description{
For the specified date range and US states, downloads AirNow, AirSIS, or WRCC data from
the AirFire archive.
}
\examples{
dt1 <- as.Date("2018-11-01") dt2 <- as.Date("2018-11-30") 
an_ws <- get_monitor_daterange(dt1, dt2, c("CA", "NV"), "airnow")
}
