% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{get_purpleair_daterange}
\alias{get_purpleair_daterange}
\title{get_purpleair_daterange}
\usage{
get_purpleair_daterange(start, end, states)
}
\arguments{
\item{start}{Date The earliest date to search.}

\item{end}{Date The latest date to search.}

\item{states}{character The states to include as a vector of two-character
state abbreviations}
}
\value{
A \emph{pas} object of outdoor PurpleAir data from the requested
  location and time period
}
\description{
get_purpleair_daterange
}
\examples{
pa_data <- get_purpleair_daterange(dt1, dt2, "CA")
}
