% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrrr-smoke.R
\name{hrrr_acquire}
\alias{hrrr_acquire}
\title{Title}
\usage{
hrrr_acquire(
  date,
  outpath,
  extent_grid,
  tz = "America/Los_Angeles",
  crs = "EPSG:3395",
  delete_files = FALSE
)
}
\arguments{
\item{date}{}

\item{outpath}{}

\item{extent_grid}{}

\item{tz}{}

\item{crs}{}

\item{delete_files}{}
}
\description{
Title
}
