% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrrr-smoke.R
\name{hrrr_at_sites}
\alias{hrrr_at_sites}
\title{hrrrr_at_sites}
\usage{
hrrr_at_sites(mon, input_path = "./data/hrrr/processed/")
}
\arguments{
\item{mon}{A SpatialPointsDataFrame with monitor data such as from
\code{\link{recast_monitors}}}
}
\value{
The data frame from \emph{mon} with the extracted values from the HRRR-smoke
  data appended
}
\description{
Extract smoke PM2.5, surface temperature, winds, RF, precip, and planetary boundary
layer height from pre-downloaded and processed HRRRR-smoke data
}
\examples{
hrrrr <- hrrr_at_sites(mon)
}
