% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{krige_airnow}
\alias{krige_airnow}
\title{krige_airnow}
\usage{
krige_airnow(an_data, vgms)
}
\arguments{
\item{an_data}{Named list from \code{\link{split_airnow_data}}.}

\item{vgms}{A list of daily variograms from
\code{\link{create_airnow_variograms}}.}
}
\value{
A data frame of the test data with the kriged values of PM25_log from
  AirNow and/or AIRSIS data.
}
\description{
Run ordinary kriging interpolation on the training data at the test locations
using the data set created by \code{\link{split_airnow_data}} and the
variograms created by \code{\link{create_airnow_variograms}}. This is to
create subset variograms that can be used in developing and validating the
final RF model. To interpolate at all sites, use
\code{\link{krige_airnow_all}}.
}
\examples{
mon_split <- split_airnow_data(an_ws, test_fraction = 0.3)
          an_vg <- create_airnow_variograms(an_ws)
          an_ok <- krige_airnow(mon_split, an_vg)
}
