% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{krige_airnow_all}
\alias{krige_airnow_all}
\title{krige_airnow_all}
\usage{
krige_airnow_all(an, outlocs, vgms)
}
\arguments{
\item{an}{A SpatialPointsDataFrame from \code{\link{recast_monitors}}. These
are the input locations with date specific data to be interpolated.}

\item{outlocs}{A SpatialPointsDataFrame from \code{\link{recast_monitors}}.
These are the output locations to be estimated.}

\item{vgms}{A list of daily variograms from
\code{\link{create_airnow_variograms}}.}
}
\value{
The dataframe portion of \emph{outlocs} with the kriging prediction
  and variance added (PM25_log_ANK and PM25_log_var)
}
\description{
Runs daily ordinary kriging based on PM25_log data from a monitor SPDF from
\code{\link{recast_monitors}} to a set of locations and dates.
}
\examples{
ank_complete <- krige_airnow_all(an_ws, an_ws an_vg)
}
