% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{krige_airnow_sitedates}
\alias{krige_airnow_sitedates}
\title{krige_airnow_sitedates}
\usage{
krige_airnow_sitedates(an, outlocs, vgms)
}
\arguments{
\item{an}{A SpatialPointsDataFrame as produced by
\code{\link{recast_monitors}}}

\item{outlocs}{A SpatialPointsDataFrame of locations (and dates) to predict}

\item{vgms}{Daily variograms as produced by
\code{\link{create_airnow_variograms}}}
}
\value{
A dataframe with the data from \emph{outlocs} with log PM2.5
  (PM25_log_ANK) and variability (PM25_log_var) appended
}
\description{
Krige monitor data for all locations and dates in an input
SpatialPointsDataFrame
}
\examples{
ank <- krige_airnow_sitedates(mon, locations, an_vg)
}
