% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{krige_purpleair}
\alias{krige_purpleair}
\title{krige_purpleair}
\usage{
krige_purpleair(pa_data, out_locs, vgms)
}
\arguments{
\item{pa_data}{Cleaned PurpleAir data such as produced by
\code{\link{purpleair_clean_spatial_outliers}}}

\item{out_locs}{A SpatialPointsDataFrame with monitor data such as from
\code{\link{recast_monitors}}. This provides the locations and dates that
will be interpolated.}

\item{vgms}{Daily variograms as produced by
\code{\link{create_purpleair_variogram}}}
}
\value{
The data from \emph{out_locs} with interpolated PurpleAir PM2.5 in
  log scale attached
}
\description{
Performs daily ordinary kriging interpolation at specified locations and
dates
}
\examples{
pa_ok <- krige_purpleair(pa_clean, mon, pa_vg)
}
