% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{krige_purpleair_sitedates}
\alias{krige_purpleair_sitedates}
\title{krige_purpleair_sitedates}
\usage{
krige_purpleair_sitedates(pa_data, outlocs, vgms)
}
\arguments{
\item{pa_data}{PurpleAir data as created by
\code{\link{get_purpleair_daterange}} or
\code{\link{create_purpleair_archive}}}

\item{outlocs}{A SpatialPointsDataFrame of locations (and dates) to predict}

\item{vgms}{Daily variograms produced by
\code{\link{create_purpleair_variograms}}}
}
\value{
A dataframe with the data from \emph{outlocs} with log PM2.5
  (PM25_log_PAK) and variability (PM25_log_PAvar) appended
}
\description{
Krige PurpleAir data for all locations and dates in a SpatialPointsDataFrame
}
\examples{
pa_ok <- krige_purpleair_sitedates(pa_clean, locations, pa_vg)
}
