% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac.R
\name{maiac_at_airnow}
\alias{maiac_at_airnow}
\title{maiac_at_airnow}
\usage{
maiac_at_airnow(
  an,
  maiac_path = "./data/MAIAC/",
  tiles_needed = c("h08v04", "h08v05", "h09v04")
)
}
\arguments{
\item{an}{A SpatialPointsDataFrame with monitor data such as from
\code{\link{recast_monitors}}}

\item{maiac_path}{The path to the MAIAC data (defaults to "./data/MAIAC/")}

\item{tiles_needed}{The MODIS tiles required for the region of interest.
Default is c("h08v04", "h08v05", "h09v04") which covers California.}
}
\value{
The data frame from \emph{an} with the extracted values from the
  MAIAC data appended
}
\description{
Extract aerosol optical depth values at point locations on given dates from
pre-downloaded MAIAC aerosol tiles (MCD19A2).
}
\examples{
 maiac <- maiac_at_airnow(mon)
}
