% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac.R
\name{maiac_download}
\alias{maiac_download}
\title{maiac_download}
\usage{
maiac_download(
  dt,
  user,
  password,
  outpath = "./data/MAIAC/",
  tiles_needed = c("h08v04", "h08v05", "h09v04")
)
}
\arguments{
\item{dt}{Date to retrieve}

\item{user}{Username NASA Earthdata}

\item{password}{Password for NASA Earthdata}

\item{outpath}{output path to write files}

\item{tiles_needed}{specific MODIS tiles required for your domain. See
https://modis-land.gsfc.nasa.gov/pdf/sn_bound_10deg.txt}
}
\description{
Download MAIAC collection 6 data from the USGS archive
}
