% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_terra.R
\name{monitors_combine}
\alias{monitors_combine}
\title{Title}
\usage{
monitors_combine(
  date,
  input_path_airnow,
  input_path_temp_monitors,
  output_path = "./processed_data/monitors/",
  extent = NULL
)
}
\arguments{
\item{date}{}

\item{input_path_airnow}{}

\item{input_path_temp_monitors}{}

\item{output_path}{}

\item{extent}{}
}
\description{
Title
}
