% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_terra.R
\name{monitors_variogram}
\alias{monitors_variogram}
\title{Fit a variogram model to monitor PM2.5 data}
\usage{
monitors_variogram(mon, cutoff = NULL, width = 15000)
}
\arguments{
\item{mon}{A \code{SpatVector} of monitor data with a \code{PM25_log} column
(e.g., from \code{\link{monitors_combine}}).}

\item{cutoff}{Maximum distance (in CRS map units) for variogram estimation. If
\code{NULL}, \code{gstat} defaults to half the diagonal of the bounding box.}

\item{width}{Bin width (lag size) for the empirical variogram, in CRS map units.
Default is \code{15000} (15 km for a metric CRS).}
}
\value{
A \code{variogramModel} object fitted with a Gaussian model and nugget
  of 0.02.
}
\description{
Computes an empirical variogram from log-transformed PM2.5 values in a
\code{SpatVector} of monitors and fits a Gaussian model with a nugget of 0.02.
The fitted model is intended for use with \code{\link{monitors_krige_grid}} or
\code{\link{monitors_krige_points}}.
}
\examples{
\dontrun{
mon <- monitors_combine("2024-11-15", input_path_airnow = "./processed_data/airnow/",
                        input_path_temp_monitors = "./processed_data/temp_monitors/")
vgm <- monitors_variogram(mon)
}
}
