% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narr.R
\name{narr_at_airnow}
\alias{narr_at_airnow}
\title{narr_at_airnow}
\usage{
narr_at_airnow(an_data)
}
\arguments{
\item{an_data}{A SpatialPointsDataFrame with monitor data such as from
\code{\link{recast_monitors}}}
}
\value{
The data frame from \emph{an_data} with the extracted values from the
  MAIAC data appended
}
\description{
Extract surface temperature, winds, RF, precip, and planetary boundary layer
height from pre-downloaded NARR data
}
\examples{
narr <- narr_at_airnow(mon)
}
