% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq.R
\name{openaq_get_averages}
\alias{openaq_get_averages}
\title{For the sites found with \code{\link{openaq_find_sites}}, download daily
average PM2.5 data from the OpenAQ API}
\usage{
openaq_get_averages(sites, date_from, date_to)
}
\arguments{
\item{sites}{A data frame returned by \code{\link{openaq_find_sites}}. The
relevant fields are id, which is the OpenAQ location id, plus latitude and
longitude}

\item{date_from}{Earliest date to query in character format "YY-mm-dd"}

\item{date_to}{Latest date to query in character format "YY-mm-dd"}
}
\value{
A data frame with daily avereage PM2.5
}
\description{
For the sites found with \code{\link{openaq_find_sites}}, download daily
average PM2.5 data from the OpenAQ API
}
\examples{
ca <- sf::st_read("states_hi_res.shp")
          ca <- dplyr::filter(ca, STATE_NAME == "California")
          ca <- dplyr::select(ca, OBJECTID, STATE_NAME)
          proj_string <- "+proj=aea +lat_1=30 +lat_2=50 +lat_0=40 +lon_0=-125 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs "
          ca_ea <- sf::st_transform(ca, crs = proj_string)
          sites <- openaq_find_sites(ca_ea)
          date_from <- "2021-02-01"
          date_to <- "2021-02-28"
          avgs <- openaq_get_averages(sites, date_from, date_to)
}
