% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_batch_download_sensors}
\alias{pa_batch_download_sensors}
\title{Download all of the PurpleAir sensors within a bounding polygon and a time
range, one sensor at a time. This requires a PurpleAir API read key with
access to historical data. Note that the PurpleAir API has a nominal limit of
1000 queries per key per day.}
\usage{
pa_batch_download_sensors(
  first_date,
  last_date,
  bounding_poly,
  api_key,
  output_folder,
  starting_point = NULL
)
}
\arguments{
\item{first_date}{Date First day of data to request}

\item{last_date}{Date Last day of data to request}

\item{bounding_poly}{A simple features polygon, such as a state or country}

\item{api_key}{A PurpleAir API read key with access to historical data}

\item{output_folder}{The location where the output RDS files should be stored}

\item{starting_point}{If specified, the rownumber to begin processing.}
}
\description{
Download all of the PurpleAir sensors within a bounding polygon and a time
range, one sensor at a time. This requires a PurpleAir API read key with
access to historical data. Note that the PurpleAir API has a nominal limit of
1000 queries per key per day.
}
