% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_build_dataset}
\alias{pa_build_dataset}
\title{Compile downloaded individual sensor information from the PurpleAir API into
a single data set covering a specified time period}
\usage{
pa_build_dataset(
  start_date,
  end_date,
  bounding_poly,
  api_key,
  sensor_folder,
  sensors = NULL
)
}
\arguments{
\item{start_date}{Date The first date of the time period}

\item{end_date}{Date The last date of the time period}

\item{sensor_folder}{The folder location of the sensor data in individual
files acquired from \code{\link{pa_batch_download_sensors}}}

\item{sensors}{A data frame from \code{\link{pa_find_sensors}}}
}
\value{
A data frame of PurpleAir data that can be passed to other routines,
  specifically \code{\link{purpleair_spatial}}
}
\description{
Compile downloaded individual sensor information from the PurpleAir API into
a single data set covering a specified time period
}
