% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_sensor_history}
\alias{pa_sensor_history}
\title{Get the historical data for a specific PurpleAir sensor over a time period.
This requires a PurpleAir API key with access to historical data, which can
be requested.}
\usage{
pa_sensor_history(sensor_index, start_date, end_date, key)
}
\arguments{
\item{sensor_index}{The numeric PurpleAir sensor index for for the sensor to request}

\item{start_date}{Date the first date to request}

\item{end_date}{Date the last date to request}

\item{key}{A PurpleAir API read key with access to historical data}
}
\description{
Get the historical data for a specific PurpleAir sensor over a time period.
This requires a PurpleAir API key with access to historical data, which can
be requested.
}
