% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_sensor_history_daily}
\alias{pa_sensor_history_daily}
\title{Get the historical data for a specific PurpleAir sensor for a single local day. This
requires a PurpleAir API key with access to historical data, which can be requested.
For multiple days, use pa_sensor_history. This function specifies the exact hours
needed in local time to avoid additional request costs.}
\usage{
pa_sensor_history_daily(
  sensor_index,
  date,
  timezone = "America/Los_Angeles",
  key
)
}
\arguments{
\item{sensor_index}{The numeric PurpleAir sensor index for for the sensor to request}

\item{date}{The date to retrieve}

\item{timezone}{The sensor local time zone (default is "America/Los_Angeles")}

\item{key}{A PurpleAir API read key with access to historical data}
}
\value{
A data frame with columns \code{sensor_index}, \code{time_stamp}, and
  \code{pm2.5_atm}, or \code{NULL} if no data is available or an API error occurs.
}
\description{
Get the historical data for a specific PurpleAir sensor for a single local day. This
requires a PurpleAir API key with access to historical data, which can be requested.
For multiple days, use pa_sensor_history. This function specifies the exact hours
needed in local time to avoid additional request costs.
}
\examples{
\dontrun{
pa_sensor_history_daily(
  sensor_index = 12345,
  date = "2024-11-15",
  key = Sys.getenv("PURPLEAIR_READ_KEY")
)
}
}
