% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_grid}
\alias{predict_grid}
\title{Title}
\usage{
predict_grid(dt, grid_file, model_file, paths)
}
\arguments{
\item{dt}{}

\item{grid_file}{}

\item{model_file}{}

\item{paths}{}
}
\description{
Title
}
