% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_locs}
\alias{predict_locs}
\title{predict_locs}
\usage{
predict_locs(
  dt1,
  dt2,
  states = "CA",
  model,
  locations,
  pa_cutoff = 1e+05,
  bluesky_special = NULL,
  pa_data = NULL
)
}
\arguments{
\item{dt1}{Date The earliest date of data to predict}

\item{dt2}{Date The latest date to predict}

\item{states}{character A vector of two-character state codes}

\item{model}{A final model object, such as extracted from the results of
\code{\link{develop_model}}. For example, if the object was called mod,
this would be mod$model$finalModel}

\item{locations}{A SpatialPointsDataFrame of locations (and dates) to predict}

\item{pa_cutoff}{A cutoff value passed to \code{\link[gstat]{vgm}} for
interploation of PurpleAir data}

\item{bluesky_special}{character Handles three special cases of BlueSky
data. If "2020", processes data prior to October 10, 2020 separately from
that after, as the BlueSky data format changed. If "HAQAST", uses custom
BlueSky-CMAQ output created during the HAQAST campaign (see
\url{https://doi.org/10.1080/10962247.2021.1891994}{O'Neill et al., 2021}).
If "nominal", instead use a nominal placeholder value for BlueSky PM2.5 of
0.1, which will have a near neutral impact on predictions.}

\item{pa_data}{Optional PurpleAir data in pas-like format. If not provided, a
download will be attempted.}
}
\value{
A dataframe with all model input values and the resulting predictions
  for the locations and dates specified in \emph{locations}
}
\description{
Uses an existing model, such as developed by \code{\link{develop_model}}, to
predict values a given locations and dates. Prepares the data necessary to do
so. MAIAC, NARR, BlueSky, and (optionally) PurpleAir data should already be
downloaded.
}
\examples{
pred <- predict_locs(dt1, dt2, states = "CA", model = mod$model$finalModel,
                               locations = locs, pa_data = pa)
}
