% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_locs_h}
\alias{predict_locs_h}
\title{Title}
\usage{
predict_locs_h(
  dt1,
  dt2,
  states = "CA",
  model,
  locations,
  pa_cutoff = 1e+05,
  pa_data = NULL,
  hrrr_path = "./data/hrrr/processed/",
  maiac_path = "./data/MAIAC/C61/"
)
}
\arguments{
\item{dt1}{}

\item{dt2}{}

\item{states}{}

\item{model}{}

\item{locations}{}

\item{pa_cutoff}{}

\item{pa_data}{}

\item{hrrr_path}{}
}
\description{
Title
}
