% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{preprocessed_bluesky_at_airnow}
\alias{preprocessed_bluesky_at_airnow}
\title{preprocessed_bluesky_at_airnow}
\usage{
preprocessed_bluesky_at_airnow(stack, an)
}
\arguments{
\item{stack}{A raster stack with daily 24-hr average PM2.5 values, such as
from \code{\link{stack_bluesky_archive}}}

\item{an}{A SpatialPointsDataFrame with monitor data such as from
\code{\link{recast_monitors}}}
}
\value{
The data frame from \emph{an} with the extracted values from the
  bluesky raster stack appended
}
\description{
Take a raster stack of 24-hr avg PM2.5 from
\code{\link{stack_bluesky_archive}} or \code{\link{stack_haqast_archive}} and
extract values at locations and dates from \code{\link{recast_monitors}}.
}
\examples{
bluesky <- preprocessed_bluesky_at_airnow(bluesky_stack, mon)
}
