% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{purpleair_clean_spatial_outliers}
\alias{purpleair_clean_spatial_outliers}
\title{purpleair_clean_spatial_outliers}
\usage{
purpleair_clean_spatial_outliers(spdf)
}
\arguments{
\item{spdf}{A SpatialPointsDataFrame of PurpleAir data such as from
\code{\link{purpleair_spatial}}}
}
\value{
The data from \emph{spdf} with spatial outliers removed
}
\description{
Scrub spatial outliers by removing and sensor that is >2 sd away from the
local median within a 10-km radius
}
\examples{
    pa_data <- get_purpleair_daterange(dt1, dt2, states)
pas <- purpleair_spatial(pa_data)
pa_clean <- purpleair_clean_spatial_outliers(pas)

}
