% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{purpleair_spatial}
\alias{purpleair_spatial}
\title{purpleair_spatial}
\usage{
purpleair_spatial(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Get log of daily mean, select only one sensor per location, convert to
SpatialPointsDataFrame, and project to planar coordinates
}
