% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidfire-package.R
\docType{package}
\name{rapidfire-package}
\alias{rapidfire}
\alias{rapidfire-package}
\title{rapidfire: relatively accurate particulate information derived from inputs retrieved easily}
\description{
Tools for acquiring input data, developing regression models, and using those models to estimate surface level PM2.5 concentrations, especially during wildfire smoke episodes.
}
\keyword{internal}
