% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{recast_monitors}
\alias{recast_monitors}
\title{recast_monitors}
\usage{
recast_monitors(mon)
}
\arguments{
\item{mon}{ws_monitor object of AirNow or AIRSIS data}
}
\value{
A SpatialPointsDataFrame
}
\description{
Switch from the compact \emph{ws_monitor} format to a square tibble, get
daily mean, calculate Log, convert to spatial, and reproject to planar
coordinates
}
\examples{
dt1 <- as.Date("2018-11-01")
  dt2 <- as.Date("2018-11-30")
  an_ws <- get_airsis_daterange(dt1, dt2, c("CA", "NV")) \%>\%
    recast_monitors()
}
