% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{split_airnow_data}
\alias{split_airnow_data}
\title{split_airnow_data}
\usage{
split_airnow_data(df, test_fraction = 0.3, seed = 1977)
}
\arguments{
\item{df}{A SpatialPointsDataFrame from \code{\link{recast_monitors}}. Can
include AirNow and/or AIRSIS data.}

\item{test_fraction}{The fraction of data to withold as the test data set.
Default is 0.3.}

\item{seed}{numeric A starting randomization seed. Default is 1977.}
}
\value{
A list with two SpatialPointsDataFrames, test and train
}
\description{
Create a training and test data set - test dataset is a fraction of the
samples, chosen randomly for each day.
}
\examples{
mon_split <- split_airnow_data(an_ws, test_fraction = 0.3)
}
