% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{stack_bluesky_archive}
\alias{stack_bluesky_archive}
\title{stack_bluesky_archive}
\usage{
stack_bluesky_archive(dt1, dt2, path = "./data/bluesky/archive")
}
\arguments{
\item{dt1}{Date of the first bluesky raster}

\item{dt2}{Date of the last bluesky raster}

\item{path}{character path to the archive files (default is
"./data/bluesky/archive")}
}
\value{
a raster stack of daily bluesky data
}
\description{
Take all of the bluesky archive files from a given path within a date range
and read them into a raster stack. All days within the range should be
available.
}
\examples{
bluesky_stack <- stack_bluesky_archive(dt1, dt2)
}
