% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{stack_haqast_archive}
\alias{stack_haqast_archive}
\title{stack_haqast_archive}
\usage{
stack_haqast_archive(dt1, dt2, path = "./data/bluesky/HAQAST_S1")
}
\arguments{
\item{dt1}{Date of the first bluesky raster}

\item{dt2}{Date of the last bluesky raster}

\item{path}{character path to the archive files (default is
"./data/bluesky/HAQAST_S1")}
}
\value{
a raster stack of daily HAQAST bluesky data
}
\description{
See \code{\link{stack_bluesky_archive}}. Works in the same way, but uses the
CMAQ output produced by the HAQAST Northern California Wildfires Tiger Team
project.
}
\examples{
bluesky_stack <- stack_haqast_archive(dt1, dt2)
}
