% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow.R
\name{airnow_acquire}
\alias{airnow_acquire}
\title{Title}
\usage{
airnow_acquire(
  date,
  output_path = "./processed_data/airnow/",
  crs = "EPSG:3395"
)
}
\arguments{
\item{date}{}

\item{output_path}{}

\item{crs}{}
}
\description{
Title
}
