% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac.R
\name{maiac_acquire}
\alias{maiac_acquire}
\title{Download MAIAC AOD data from NASA EartData archive and save it as tif}
\usage{
maiac_acquire(
  dt,
  user = NULL,
  password = NULL,
  outpath = "./raw_data/MAIAC/",
  bounding_box = c(-124.5, 32.5, -114, 42.1),
  tz = "America/Los_Angeles"
)
}
\arguments{
\item{dt}{character or date - date of MAIAC data to acquire}

\item{user}{username for NASA EarthData login}

\item{password}{password for NASA EartData login}

\item{outpath}{output path for files}

\item{bounding_box}{a vector of coordinates, x1, y1, x2, y2 to bound the request. The
default bounding box covers California}

\item{tz}{time zone for the requested location (default is America/Los_Angeles)}
}
\value{
Returns the list of tif files and downloads files to disk
}
\description{
Download MAIAC AOD data from NASA EartData archive and save it as tif
}
\examples{
maiac_acquire("2025-09-05", user = u, password = p)
}
