% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac.R
\name{maiac_preprocess}
\alias{maiac_preprocess}
\title{Title}
\usage{
maiac_preprocess(
  dt,
  input_path = "./raw_data/MAIAC/",
  output_path = "./processed_data/MAIAC/",
  overwrite = FALSE
)
}
\arguments{
\item{dt}{}

\item{input_path}{}

\item{output_path}{}

\item{overwrite}{}
}
\description{
Title
}
