% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_terra.R
\name{monitors_variogram}
\alias{monitors_variogram}
\title{Title}
\usage{
monitors_variogram(mon, cutoff = NULL, width = 15000)
}
\arguments{
\item{mon}{}

\item{cutoff}{}

\item{width}{}
}
\description{
Title
}
