% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_acquire_realtime}
\alias{pa_acquire_realtime}
\title{Title}
\usage{
pa_acquire_realtime(
  nwlng = -124.41,
  nwlat = 42.01,
  selng = -114.13,
  selat = 32.53,
  max_age = 3600,
  key = NULL
)
}
\arguments{
\item{nwlng}{}

\item{nwlat}{}

\item{selng}{}

\item{selat}{}

\item{max_age}{}

\item{key}{}
}
\description{
Title
}
