% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_check_api_key}
\alias{pa_check_api_key}
\title{Check that your PurpleAir API key is working and that it is the correct type}
\usage{
pa_check_api_key(key)
}
\arguments{
\item{key}{A PurpleAir API key}
}
\value{
The type of API key. Either "READ" or "WRITE"
}
\description{
Check that your PurpleAir API key is working and that it is the correct type
}
