% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_find_sensors}
\alias{pa_find_sensors}
\title{Get all of the PurpleAir sensor indices within a bounding box. See
https://api.purpleair.com/#api-sensors-get-sensors-data}
\usage{
pa_find_sensors(nwlng, nwlat, selng, selat, key)
}
\arguments{
\item{nwlng}{A northwest longitude for the bounding box}

\item{nwlat}{A northwest latitude for the bounding box}

\item{selng}{A southeast longitude for the bounding box}

\item{selat}{A southeast latitude for the bounding box}

\item{key}{A PurpleAir READ API key}
}
\value{
A data frame with sensor_index, date_created, last_seen, name,
  latitude, and longitude for all sensors within the specified bounding box.
}
\description{
Get all of the PurpleAir sensor indices within a bounding box. See
https://api.purpleair.com/#api-sensors-get-sensors-data
}
\examples{
# CA Bounding Box
nwlng <- -124.41
nwlat <- 42.01
selng <- -114.13
selat <- 32.53
pa_find_sensors(nwlng, nwlat, selng, selat, key = my_api_key)
}
