% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair.R
\name{pa_preprocess}
\alias{pa_preprocess}
\title{Take realtime purpleair data from a duckdb file and create 24-hr averages ready for
interpolation}
\usage{
pa_preprocess(
  dt,
  duckdb_path,
  sensors,
  location_types = 0,
  timezone = "America/Los_Angeles",
  clean_outliers = TRUE,
  crs = "EPSG:3395",
  output_path = "./processed_data/purpleair/"
)
}
\arguments{
\item{output_path}{}
}
\description{
Take realtime purpleair data from a duckdb file and create 24-hr averages ready for
interpolation
}
