% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_monitors.R
\name{temp_monitors_acquire}
\alias{temp_monitors_acquire}
\title{Title}
\usage{
temp_monitors_acquire(
  date,
  states = NULL,
  timezone = "America/Los_Angeles",
  output_path = "./processed_data/temp_monitors/",
  crs = "EPSG:3395"
)
}
\arguments{
\item{date}{}

\item{states}{}

\item{timezone}{}

\item{output_path}{}

\item{crs}{}
}
\description{
Title
}
